Add RangeStream to _write_block for ASDF
Added the use of a RangeStream for the _write_block function in the
AsdfWriter. This allows to write ASDF files given a list of blocks. For
some reason it would not update the offset and keep seeking to the same
offset within the given range, using the RangeStream as a source for
shutil.copyfileobj fixes this.