Fix argument for shutil.copyobj
After digging into it somewhat deeper it turns out that the way we use
`shutil.copyobj` was wrong and the `size` argument was interpreted
differently. Using the `RangeStream` still fixes the issue that we only
read a specified length of bytes, whilst before the `size` argument was
interpreted as the size that the buffer should have after reading whilst
this is actually the block size for reading the bytes.

Removed the `size` as an argument to `shutil.copyobj`, using
`RangeStream` to make sure we read the specified amount of bytes in
`size`.